refactor: reorganize widgets and centralize colors
- Move streaming_indicator_widget.dart from tests/lib to src/client/lib
- Move message_bubble_widget.dart from tests/lib to src/client/lib
- All 3 presentation widgets now import colors from centralized AppColors
- Replace deprecated Color.withOpacity() with Color.withValues(alpha:)
- Update test imports to use package:softarchitect_ai instead of package:softarchitect_ai_tests
- All 20 widget tests passing
- Zero compilation errors or warnings
