+++ src/client/lib/features/chat/presentation/notifiers/chat_notifier.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_riverpod/legacy.dart';
 
 import '../../domain/entities/chat_message.dart';
 import '../../domain/entities/document_proposal.dart';
